pyproject.toml
[tool.poetry]
name = "langchain agent application"
version = "0.1.0"
description = ""
authors = ["Michael"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
altair = "5.2.0"
annotated-types = "0.6.0"
anyio = "4.2.0"
asgiref = "3.7.2"
async-timeout = "4.0.3"
attrs = "23.2.0"
backoff = "2.2.1"
bcrypt = "4.1.2"
beautifulsoup4 = "4.12.2"
blinker = "1.7.0"
build = "1.0.3"
cachetools = "5.3.2"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
click = "8.1.7"
colorama = "0.4.6"
coloredlogs = "15.0.1"
dataclasses-json = "0.6.3"
deprecated = "1.2.14"
distro = "1.9.0"
exceptiongroup = "1.2.0"
fastapi = "0.109.0"
filelock = "3.13.1"
flatbuffers = "23.5.26"
frozenlist = "1.4.1"
fsspec = "2023.12.2"
gitdb = "4.0.11"
gitpython = "3.1.41"
google-auth = "2.26.2"
google-search-results = "2.4.2"
googleapis-common-protos = "1.62.0"
greenlet = "3.0.3"
grpcio = "1.60.0"
h11 = "0.14.0"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.25.2"
huggingface-hub = "0.20.3"
humanfriendly = "10.0"
idna = "3.6"
importlib-metadata = "6.11.0"
importlib-resources = "6.1.1"
jinja2 = "3.1.3"
joblib = "1.3.2"
jsonpatch = "1.33"
jsonpointer = "2.4"
jsonschema = "4.21.1"
jsonschema-specifications = "2023.12.1"
kubernetes = "29.0.0"
langchain-openai = "0.0.2.post1"
langchainhub = "0.1.14"
lark = "1.1.9"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.20.2"
mdurl = "0.1.2"
mmh3 = "4.1.0"
monotonic = "1.6"
mpmath = "1.3.0"
mypy-extensions = "1.0.0"
networkx = "3.2.1"
nltk = "3.8.1"
numexpr = "2.9.0"
numpy = "1.26.3"
oauthlib = "3.2.2"
ollama = "0.1.6"
openai = "1.7.2"
opentelemetry-api = "1.22.0"
opentelemetry-exporter-otlp-proto-common = "1.22.0"
opentelemetry-exporter-otlp-proto-grpc = "1.22.0"
opentelemetry-instrumentation = "0.43b0"
opentelemetry-instrumentation-asgi = "0.43b0"
opentelemetry-instrumentation-fastapi = "0.43b0"
opentelemetry-proto = "1.22.0"
opentelemetry-sdk = "1.22.0"
opentelemetry-semantic-conventions = "0.43b0"
opentelemetry-util-http = "0.43b0"
overrides = "7.4.0"
packaging = "23.2"
pandas = "2.2.0"
pillow = "10.2.0"
posthog = "3.3.1"
protobuf = "4.25.2"
pulsar-client = "3.4.0"
pyarrow = "14.0.2"
pyasn1 = "0.5.1"
pyasn1-modules = "0.3.0"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pydeck = "0.8.1b0"
pygments = "2.17.2"
pypdf = "3.17.4"
pypika = "0.48.9"
pyproject-hooks = "1.0.0"
pyreadline3 = "3.4.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
rank-bm25 = "0.2.2"
referencing = "0.32.1"
regex = "2023.12.25"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
rich = "13.7.0"
rpds-py = "0.17.1"
rsa = "4.9"
scikit-learn = "1.3.2"
scipy = "1.11.4"
sentence-transformers = "2.2.2"
simsimd = "3.6.3"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.0"
soupsieve = "2.5"
sqlalchemy = "2.0.25"
starlette = "0.35.1"
streamlit = "1.30.0"
sympy = "1.12"
tenacity = "8.2.3"
text-generation = "0.6.1"
threadpoolctl = "3.2.0"
tiktoken = "0.5.2"
toml = "0.10.2"
tomli = "2.0.1"
toolz = "0.12.0"
tornado = "6.4"
tqdm = "4.66.1"
transformers = "4.37.2"
typer = "0.9.0"
types-requests = "2.31.0.20240125"
typing-inspect = "0.9.0"
typing-extensions = "4.9.0"
tzdata = "2023.4"
tzlocal = "5.2"
urllib3 = "2.1.0"
uvicorn = "0.25.0"
validators = "0.22.0"
watchdog = "3.0.0"
watchfiles = "0.21.0"
websocket-client = "1.7.0"
websockets = "12.0"
wrapt = "1.16.0"
yarl = "1.9.4"
zipp = "3.17.0"
astrapy = "0.7.6"
cassandra-driver = "3.29.0"
cassio = "0.1.5"
contourpy = "1.2.0"
cycler = "0.12.1"
deprecation = "2.1.0"
fonttools = "4.50.0"
geomet = "0.2.1.post1"
h2 = "4.1.0"
hpack = "4.0.0"
hyperframe = "6.0.1"
kiwisolver = "1.4.5"
langchain-astradb = "0.0.1"
matplotlib = "3.8.3"
orjson = "3.10.0"
psycopg2 = "2.9.9"
pyparsing = "3.1.2"
streamlit-option-menu = "0.3.12"
langgraph = "^0.0.44"
langchain = "^0.1.17"
langchain-experimental = "^0.0.57"
chromadb = "^0.5.0"
pinecone = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
